+++ reserver/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Reserver modules."""
 from .reserver_param import RESERVER_VERSION
-from .reserver_obj import Uploader
+from .reserver_obj import PyPIUploader
 
 __version__ = RESERVER_VERSION