+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `batch_upload` method added to `PyPIUploader`
 ### Changed
+- `Uploader` changed to `PyPIUploader`
 - `README.md` modified
 ## [0.1] - 2024-02-07
 ### Added